prep: remove the SDL3 traps that would not fail to compile
Groundwork for an SDL3 port, all of it correct and useful on SDL2 today.
These are the changes that would otherwise migrate silently and wrongly.

Timers move to 64 bits, read through SDL_GetTicks64 (SDL >= 2.0.18, already
required, and used elsewhere in the tree). This is a real fix, not only
preparation: SDL_GetTicks() wraps after 49.7 days and NTimer's
"starttime + timeout" wrapped with it, so on a server up that long
isTimeOut() began returning true immediately and stayed that way. NTimer
backs socket timeouts, bot timers, player state and the input processor. The
frame limiter, the heartbeat and the font cache's age sweep get the same
treatment. TimeStamp deliberately stays 32-bit -- it is the wire format, and
getNetworkTicks() says so.

Init results are compared explicitly. SDL2 returns 0 from SDL_Init and
SDL_InitSubSystem; SDL3 returns bool, where true means success. So
"if (SDL_InitSubSystem(SDL_INIT_VIDEO))" would throw on success after the
upgrade, and "SDL_InitSubSystem(SDL_INIT_AUDIO) < 0" would become
permanently false and swallow the failure. Both compile fine either way,
which is what makes them worth spelling out now.

Palette access goes through 2D/SurfacePalette.hpp. SDL3 makes
SDL_Surface::format an enumerated value rather than a pointer to a structure,
and indexed surfaces no longer get a palette implicitly, so the four places
reaching through surface->format->palette become one edit later instead of a
hunt through the drawing code.

Verified: 8/8 tests, clean under AddressSanitizer, server with five bots at
0.030 ms/tick and the client at 0.218 ms -- both unchanged.

+++ src/Lib/2D/CachedFontRenderer.hpp
@@ -30,16 +30,16 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 class RenderedText {
  public:
   SDL_Surface* sdlSurface;
-  Uint32 lastUsedTick;
+  Uint64 lastUsedTick;
 
   RenderedText() : sdlSurface(nullptr), lastUsedTick(0) {}
-  RenderedText(SDL_Surface* surface, Uint32 tick)
+  RenderedText(SDL_Surface* surface, Uint64 tick)
       : sdlSurface(surface), lastUsedTick(tick) {}
 };
 
 class CachedFontRenderer {
  private:
-  static Uint32 lastCleanedTick;
+  static Uint64 lastCleanedTick;
   static std::unordered_map<std::string, RenderedText> rendered_surfaces;
   static std::string create_cache_key(const char* text, SDL_Color color, SDL_Color backColor, bool wrapped, int wrapLength);
   static TTF_Font* font;
